fix(demo): update the systemjs demo to be compliant with the new rc5 version
