+++ demo/systemjs-angularquickstart-structure/package.json
@@ -13,8 +13,8 @@
     "typescript": "^1.7.5"
   },
   "dependencies": {
-    "angular2": "^2.0.0-beta.6",
-    "angular2-toaster": "0.2.0-beta.0",
+    "angular2": "^2.0.0-rc.5",
+    "angular2-toaster": "0.3.6-rc.5",
     "es6-shim": "0.33.3",
     "http-server": "^0.8.5",
     "reflect-metadata": "0.1.2",
