+++ demo/systemjs/package.json
@@ -18,13 +18,13 @@
     "rimraf": "^2.5.2"
   },
   "dependencies": {
-    "@angular/common": "2.0.0-rc.4",
-    "@angular/compiler": "2.0.0-rc.4",
-    "@angular/core": "2.0.0-rc.4",
-    "@angular/platform-browser": "2.0.0-rc.4",
-    "@angular/platform-browser-dynamic": "2.0.0-rc.4",
+    "@angular/common": "2.0.0-rc.5",
+    "@angular/compiler": "2.0.0-rc.5",
+    "@angular/core": "2.0.0-rc.5",
+    "@angular/platform-browser": "2.0.0-rc.5",
+    "@angular/platform-browser-dynamic": "2.0.0-rc.5",
     "rxjs": "5.0.0-beta.6",
-    "angular2-toaster": "0.3.6-rc.3",
+    "angular2-toaster": "0.3.6-rc.5",
     "es6-shim": "0.35.0",
     "http-server": "^0.8.5",
     "reflect-metadata": "0.1.3",
