+++ demo/systemjs/src/test.component.ts
@@ -1,0 +1,8 @@
+import {Component} from '@angular/core';
+
+@Component({
+    selector: 'test-component',
+    template: `<div>loaded via component</div>`
+})
+export class TestComponent { }
+
