+++ demo/systemjs/src/test2.component.ts
@@ -1,0 +1,8 @@
+import {Component} from '@angular/core';
+
+@Component({
+    selector: 'test-component2',
+    template: `<div>loaded via component 2</div>`
+})
+export class TestComponent2 { }
+
