+++ demo/webpack/package.json
@@ -21,9 +21,9 @@
     "systemjs": "^0.19.27",
     "zone.js": "0.6.12",
     "ts-loader": "0.8.2",
-    "@angular/common": "2.0.0-rc.1",
-    "@angular/core": "2.0.0-rc.1",
-    "@angular/platform-browser-dynamic": "2.0.0-rc.1",
-    "angular2-toaster": "0.3.0-rc.0"
+    "@angular/common": "2.0.0-rc.5",
+    "@angular/core": "2.0.0-rc.5",
+    "@angular/platform-browser-dynamic": "2.0.0-rc.5",
+    "angular2-toaster": "0.3.0-rc.5"
   }  
 }
